Refactoring
- Remove references to entity.
